+++ feature_listen_later_ui/src/main/java/com/allsoftdroid/audiobook/feature_listen_later_ui/di/FeatureListenLaterModule.kt
@@ -1,6 +1,7 @@
-package com.allsoftdroid.audiobook.feature_listen_later_ui
+package com.allsoftdroid.audiobook.feature_listen_later_ui.di
 
 import androidx.annotation.VisibleForTesting
+import com.allsoftdroid.audiobook.feature_listen_later_ui.ListenLaterViewModel
 import com.allsoftdroid.database.common.AudioBookDatabase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -15,7 +16,8 @@ import kotlin.coroutines.CoroutineContext
 
 object FeatureListenLaterModule {
 
-    fun injectFeature() = loadFeature
+    fun injectFeature() =
+        loadFeature
 
     fun unLoadModules(){
         unloadKoinModules(
